requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
aiosqlite==0.20.0
python-dotenv==1.0.1
